refactor(MovieQuizViewController): add questionPresenter and move some logic there
