+++ MovieQuiz/Models/GameState.swift
@@ -1,8 +1,6 @@
 import Foundation
 
 struct GameState {
-    var questionsAmount = 10
     var currentQuestion: QuizQuestion?
-    var currentQuestionNumber = 0
     var currentScore = 0
 }
